Rename equalsInAnyOrder to equalInAnyOrder
Rename equalsInAnyOrder to equalInAnyOrder
