+++ Expecta/Matchers/EXPMatchers.h
@@ -1,6 +1,6 @@
 #import "EXPMatchers+beNil.h"
 #import "EXPMatchers+equal.h"
-#import "EXPMatchers+equalsInAnyOrder.h"
+#import "EXPMatchers+equalInAnyOrder.h"
 #import "EXPMatchers+beInstanceOf.h"
 #import "EXPMatchers+beKindOf.h"
 #import "EXPMatchers+beSubclassOf.h"
